life-coach-agent/pyproject.toml
[project]
name = "life-coach-agent"
version = "0.1.0"
description = "OpenAI Agents SDK + Clean Architecture 기반 라이프 코치 에이전트"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.6",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.0",
    "ipykernel>=7.2.0",
]
